+++ crypto/keys/ed25519/doc.go
@@ -1,0 +1,10 @@
+package ed25519
+
+/*
+This package contains a wrapper around crypto/ed22519 to make it comply with the crypto interfaces.
+
+This package employs zip215 rules. We use https://github.com/hdevalence/ed25519consensus verification function.
+
+This package works with correctly generated signatures. To read more about what this means see https://hdevalence.ca/blog/2020-10-04-its-25519am
+
+*/